Refactor: 마이페이지 기능 디렉토리 분리에 따른 리팩토링
