+++ src/main/java/com/team1/spreet/domain/user/controller/UserController.java
@@ -1,27 +1,24 @@
 package com.team1.spreet.domain.user.controller;
 
-import com.team1.spreet.global.common.dto.CustomResponseBody;
 import com.team1.spreet.domain.user.dto.UserDto;
-import com.team1.spreet.global.common.model.SuccessStatusCode;
-import com.team1.spreet.global.auth.security.UserDetailsImpl;
 import com.team1.spreet.domain.user.service.UserService;
+import com.team1.spreet.global.auth.security.UserDetailsImpl;
+import com.team1.spreet.global.common.dto.CustomResponseBody;
+import com.team1.spreet.global.common.model.SuccessStatusCode;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
-import java.util.List;
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
 
 @Api(tags = "user")
 @RestController
@@ -59,46 +56,11 @@ public class UserController {
         return new CustomResponseBody<>(SuccessStatusCode.NICKNAME_DUPLICATE_CHECK);
     }
 
-    @ApiOperation(value = "회원정보 조회 API")
-    @GetMapping("/mypage")
-    public CustomResponseBody<UserDto.UserInfoResponseDto> getUserInfo(@AuthenticationPrincipal UserDetailsImpl userDetails) {
-        return new CustomResponseBody<>(SuccessStatusCode.GET_USER_INFO, userService.getUserInfo(userDetails.getUser()));
-    }
-
-    @ApiOperation(value = "프로필 이미지 수정 API")
-    @PutMapping("/mypage/edit/profile-image")
-    public CustomResponseBody<SuccessStatusCode> updateProfileImage(
-        @RequestParam(value = "file") @Valid @ApiParam(value = "새로운 프로필 이미지") MultipartFile file,
-        @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        userService.updateProfileImage(file, userDetails.getUser());
-        return new CustomResponseBody<>(SuccessStatusCode.UPDATE_USER_INFO);
-    }
-
-    @ApiOperation(value = "회원정보 수정 API")
-    @PutMapping("/mypage/edit/nickname")
-    public CustomResponseBody<SuccessStatusCode> updateNickname(
-        @RequestBody @Valid @ApiParam(value = "새로운 닉네임") UserDto.NicknameRequestDto requestDto,
-        @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        userService.updateNickname(requestDto, userDetails.getUser());
-        return new CustomResponseBody<>(SuccessStatusCode.UPDATE_USER_INFO);
-    }
-
-    // 회원이 작성한 게시글 목록 조회
-    @ApiOperation(value = "회원이 작성한 게시글 조회 API")
-    @GetMapping("/mypage/post")
-    public CustomResponseBody<List<UserDto.PostResponseDto>> getPostList (
-        @RequestParam(value = "classification") @ApiParam(value = "게시글 분류") String classification,
-        @RequestParam(value = "page") @ApiParam(value = "조회할 페이지") int page,
-        @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        return new CustomResponseBody<>(SuccessStatusCode.GET_POST_LIST, userService.getPostList(classification, page, userDetails.getUser()));
-    }
-
-    @ApiOperation(value = "회원 비밀번호 수정 API")
+    @ApiOperation(value = "비밀번호 재설정 API")
     @PutMapping("/reset/password")
     public CustomResponseBody<SuccessStatusCode> resetPassword(
-        @RequestBody @Valid @ApiParam(value = "수정할 비밀번호") UserDto.ResetPwRequestDto requestDto,
-        @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        userService.resetPassword(requestDto, userDetails.getUser());
+        @RequestBody @Valid @ApiParam(value = "비밀번호 재설정을 위한 정보") UserDto.ResetPwRequestDto requestDto) {
+        userService.resetPassword(requestDto);
         return new CustomResponseBody<>(SuccessStatusCode.UPDATE_PASSWORD);
     }
